requirements.txt
# general modules
django >3.2,<3.3
python-decouple
pillow
djangorestframework

# third party modules
markdown
django-filter
drf-yasg[validation]
djangorestframework-simplejwt
django-mail-templated
djoser
pytest
pytest-django
flake8
black